Move types to their own folder and change lambda names to env vars
