+++ workspaces/serverless-wrappers/types/index.ts
@@ -1,0 +1,16 @@
+export type InvokeLambdaParams = {
+  FunctionName: string;
+  InvocationType: string;
+  LogType: string;
+  Payload: string;
+};
+
+export type LambdaResponse = {
+  statusCode: number;
+  message: string;
+};
+
+export interface IRunBotOrchestratorDependencies {
+  getMessage(): LambdaResponse;
+  sendMessage(message: string): LambdaResponse;
+}
